moved get_source_xpos out of wavecorr, fixed unit test
