+++ jwst/extract_2d/nirspec.py
@@ -8,12 +8,11 @@ from gwcs.utils import _toindex
 from gwcs import wcstools
 
 from stdatamodels.jwst import datamodels
+from stdatamodels.jwst.transforms import models as trmodels
 
 from ..assign_wcs import nirspec
 from ..assign_wcs import util
 from ..lib import pipe_utils
-
-from ..wavecorr import wavecorr
 
 log = logging.getLogger(__name__)
 log.setLevel(logging.DEBUG)
@@ -86,7 +85,7 @@ def nrs_extract2d(input_model, slit_name=None):
 
             if (new_model.meta.exposure.type.lower() == 'nrs_fixedslit') and \
                 (slit.name == input_model.meta.instrument.fixed_slit):
-                wavecorr.get_source_xpos(new_model, None, None)
+                get_source_xpos(new_model)
 
             # Update the S_REGION keyword value for the extracted slit
             if 'world' in input_model.meta.wcs.available_frames:
@@ -272,3 +271,42 @@ def extract_slit(input_model, slit, exp_type):
     new_model.meta.wcs = slit_wcs
 
     return new_model, xlo, xhi, ylo, yhi
+
+
+def get_source_xpos(slit):
+    """
+    Compute the source position within the slit for a NIRSpec fixed slit.
+
+    Parameters
+    ----------
+    slit : `~jwst.datamodels.SlitModel`
+        The slit object.
+
+    Returns
+    -------
+    xpos : float
+        X coordinate of the source as a fraction of the slit size.
+    """
+    xoffset = slit.meta.dither.x_offset  # in arcsec
+    yoffset = slit.meta.dither.y_offset  # in arcsec
+    v2ref = slit.meta.wcsinfo.v2_ref  # in arcsec
+    v3ref = slit.meta.wcsinfo.v3_ref  # in arcsec
+    v3idlyangle = slit.meta.wcsinfo.v3yangle  # in deg
+    vparity = slit.meta.wcsinfo.vparity
+
+    idl2v23 = trmodels.IdealToV2V3(v3idlyangle, v2ref, v3ref, vparity)
+    log.debug("wcsinfo: {0}, {1}, {2}, {3}".format(v2ref, v3ref, v3idlyangle, vparity))
+    # Compute the location in V2,V3 [in arcsec]
+    xv, yv = idl2v23(xoffset, yoffset)
+    log.info(f'xoffset, yoffset, {xoffset}, {yoffset}')
+
+    # Position in the virtual slit
+    xpos_slit, ypos_slit, lam_slit = slit.meta.wcs.get_transform('v2v3', 'slit_frame')(
+        xv, yv, 2)
+    # Update slit.source_xpos, slit.source_ypos
+    slit.source_xpos = xpos_slit
+    slit.source_ypos = ypos_slit
+    log.debug('Source X/Y position in V2V3: {0}, {1}'.format(xv, yv))
+    log.info('Source X/Y position in the slit: {0}, {1}'.format(xpos_slit, ypos_slit))
+
+    return xpos_slit
