+++ jwst/wavecorr/tests/test_wavecorr.py
@@ -112,6 +112,15 @@ def test_skipped():
         WavecorrStep.call(im)
 
     outa = AssignWcsStep.call(im)
+
+    dither = {'x_offset': 0.0, 'y_offset': 0.0}
+    outa.meta.dither = dither
+    outa.meta.wcsinfo.v3yangle = 138.78
+    outa.meta.wcsinfo.vparity = -1
+    outa.meta.wcsinfo.v2_ref = 321.87
+    outa.meta.wcsinfo.v3_ref = -477.94
+    outa.meta.wcsinfo.roll_ref = 15.1234
+
     oute = Extract2dStep.call(outa)
     outs = SourceTypeStep.call(oute)
 
@@ -131,24 +140,16 @@ def test_skipped():
     outw = WavecorrStep.call(outs)
     assert out.meta.cal_step.wavecorr == "SKIPPED"
 
-    dither = {'x_offset': 0.0, 'y_offset': 0.0}
-    ind = np.nonzero([s.name == 'S400A1' for s in outs.slits])[0].item()
-    outs.slits[ind].meta.dither = dither
-    outs.slits[ind].meta.wcsinfo.v3yangle = 138.78
-    outs.slits[ind].meta.wcsinfo.vparity = -1
-    outs.slits[ind].meta.wcsinfo.v2_ref = 321.87
-    outs.slits[ind].meta.wcsinfo.v3_ref = -477.94
-    outs.slits[ind].meta.wcsinfo.roll_ref = 15.1234
 
     # Test step is skipped if source is "EXTENDED"
     outw = WavecorrStep.call(outs)
     assert out.meta.cal_step.wavecorr == "SKIPPED"
 
-    outs.slits[ind].source_type = 'POINT'
     outw = WavecorrStep.call(outs)
 
-    source_pos = (0.004938526981283373, -0.02795306204991911)
-    assert_allclose((outw.slits[ind].source_xpos, outw.slits[ind].source_ypos),source_pos)
+    ind = 0 # this is the only point source
+    source_pos = (33.110462,  2.329629)
+    assert_allclose((outw.slits[ind].source_xpos, outw.slits[ind].source_ypos),source_pos, rtol=1e-5)
 
     # Test if the corrected wavelengths are not monotonically increasing
 
@@ -249,8 +250,6 @@ def test_wavecorr_fs():
     assert_allclose(result.slits[0].source_xpos, 0.127111, atol=1e-6)
 
     slit = result.slits[0]
-    source_xpos = wavecorr.get_source_xpos(slit)
-    assert_allclose(result.slits[0].source_xpos, source_xpos, atol=1e-6)
 
     mean_correction = np.abs(src_result.slits[0].wavelength - result.slits[0].wavelength)
     assert_allclose(np.nanmean(mean_correction), 0.003, atol=.001)
